+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/adhearsion/blather/compare/master...develop)
+  * Feature: Add Blather::Stanza::Iq::IBR to implement XEP-0077 In-Band Registration
 
 # [v2.0.0](https://github.com/adhearsion/blather/compare/v1.2.0...v2.0.0) - [2018-06-18](https://rubygems.org/gems/blather/versions/2.0.0)
   * Bugfix: Require EventMachine >= 1.2.6 to avoid segfault issue